# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "API documentation for Shop application",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a list of orders",
                "operationId": "getOrders",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "952386297add86df04edd57d57288cd8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get details of a specific order",
                "operationId": "2264c8c11bfd386ac3cfc3ed5e792218",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an order",
                "operationId": "07963af8937ce86423453b12ee132643",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "e10a3f892d53bffd26b24df922c7f723",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginUserRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "Category": {
                "required": [
                    "id",
                    "name",
                    "description",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Category 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "category description"
                    },
                    "is_shared": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:30:00Z"
                    }
                },
                "type": "object"
            },
            "StoreCategoryRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Updated Category Name"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    }
                },
                "type": "object"
            },
            "UpdateCategoryRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Updated Category Name"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    }
                },
                "type": "object"
            },
            "UpdatePageRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Page title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Page content"
                    }
                },
                "type": "object"
            },
            "StorePageRequest": {
                "properties": {
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Page title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Page content"
                    }
                },
                "type": "object"
            },
            "Page": {
                "required": [
                    "id",
                    "author_id",
                    "title",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Page title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Page content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:30:00Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Facebook"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-12345"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "tags": {
                        "description": "List of tags associated wiht the product",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "technology"
                        }
                    }
                },
                "type": "object"
            },
            "StoreProductsRequest": {
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Facebook"
                    },
                    "description": {
                        "type": "string",
                        "example": "My Facebook password"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-12345"
                    },
                    "tags": {
                        "description": "List of tags associated wiht the product",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "technology"
                        }
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cash"
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Main St, New York, NY 10030"
                    },
                    "billing_address": {
                        "type": "string",
                        "example": "123 Main St, New York, NY 10030"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Please deliver before 5pm"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreOrderRequest": {
                "required": [
                    "user_id",
                    "status",
                    "total_price",
                    "payment_status",
                    "shipping_address",
                    "items"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cash"
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Main St, New York, NY 10030"
                    },
                    "billing_address": {
                        "type": "string",
                        "example": "123 Main St, New York, NY 10030"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Please deliver before 5pm"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "required": [
                    "product_id",
                    "quantity",
                    "price"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Mobile phone"
                    },
                    "total": {
                        "type": "decimal",
                        "example": "23,00"
                    }
                },
                "type": "object"
            },
            "SharedAccess": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "accessible_type": {
                        "type": "string",
                        "example": "CategoryModel"
                    },
                    "accessible_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreSharedAccessRequest": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "accessible_type": {
                        "type": "string",
                        "example": "CategoryModel"
                    },
                    "accessible_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "UpdateSharedAccessRequest": {
                "required": [
                    "user_id",
                    "accessible_type",
                    "accessible_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "accessible_type": {
                        "type": "string",
                        "example": "CategoryModel"
                    },
                    "accessible_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2028-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Media": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "file_name": {
                        "type": "string",
                        "example": "example.jpg"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "/uploads/example.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "EntityMedia": {
                "properties": {
                    "media_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "mediable_type": {
                        "type": "string",
                        "example": "user"
                    },
                    "mediable_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "required": [
                    "email",
                    "name",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "test#example.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "Test User"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreTagRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Tag"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "API Endpoints for managing orders"
        }
    ]
}